requirements.txt
streamlit
numpy
streamlit-code-editor
pandas
